move examples tests
