+++ test/examples/test_xss_scanner.py
